test: Consolidate connection tests
