+++ homestar-runtime/tests/network/connection.rs
@@ -1,8 +1,9 @@
 use crate::{
     make_config,
     utils::{
-        kill_homestar, listen_addr, multiaddr, wait_for_socket_connection, ChildGuard, ProcInfo,
-        TimeoutFutureExt, BIN_NAME, ED25519MULTIHASH, SECP256K1MULTIHASH,
+        check_for_line_with, kill_homestar, listen_addr, multiaddr, retrieve_output,
+        wait_for_socket_connection, ChildGuard, ProcInfo, TimeoutFutureExt, BIN_NAME,
+        ED25519MULTIHASH, SECP256K1MULTIHASH,
     },
 };
 use anyhow::Result;
@@ -75,7 +76,7 @@ fn test_connection_notifications_integration() -> Result<()> {
         .stdout(Stdio::piped())
         .spawn()
         .unwrap();
-    let _proc_guard1 = ChildGuard::new(homestar_proc1);
+    let proc_guard1 = ChildGuard::new(homestar_proc1);
 
     if wait_for_socket_connection(ws_port1, 1000).is_err() {
         panic!("Homestar server/runtime failed to start in time");
@@ -151,7 +152,7 @@ fn test_connection_notifications_integration() -> Result<()> {
             }
         }
 
-        let _ = kill_homestar(proc_guard2.take(), None);
+        let dead_proc2 = kill_homestar(proc_guard2.take(), None);
 
         // Poll for connection closed message
         loop {
@@ -167,18 +168,92 @@ fn test_connection_notifications_integration() -> Result<()> {
             }
         }
 
-        // Check node endpoint to match
-        let http_url = format!("http://localhost:{}", ws_port1);
-        let http_resp = reqwest::get(format!("{}/node", http_url)).await.unwrap();
-        assert_eq!(http_resp.status(), 200);
-        let http_resp = http_resp.json::<serde_json::Value>().await.unwrap();
-        assert_eq!(
-            http_resp,
-            serde_json::json!({
-                    "static": {"peer_id": ED25519MULTIHASH},
-                    "dynamic": {"listeners": [format!("{listen_addr1}")], "connections": {}}
-            })
+        // Kill proceses.
+        let dead_proc1 = kill_homestar(proc_guard1.take(), None);
+
+        // Retrieve logs.
+        let stdout1 = retrieve_output(dead_proc1);
+        let stdout2 = retrieve_output(dead_proc2);
+
+        // Check node one added node two to Kademlia table
+        let two_added_to_dht = check_for_line_with(
+            stdout1.clone(),
+            vec![
+                "added configured node to kademlia routing table",
+                SECP256K1MULTIHASH,
+            ],
         );
+
+        // Check node one DHT routing table was updated with node two
+        let two_in_dht_routing_table = check_for_line_with(
+            stdout1.clone(),
+            vec![
+                "kademlia routing table updated with peer",
+                SECP256K1MULTIHASH,
+            ],
+        );
+
+        // Check that node one connected to node two.
+        let one_connected_to_two = check_for_line_with(
+            stdout1.clone(),
+            vec!["peer connection established", SECP256K1MULTIHASH],
+        );
+
+        // Check that node two disconnected from node one.
+        let two_disconnected_from_one = check_for_line_with(
+            stdout1.clone(),
+            vec!["peer connection closed", SECP256K1MULTIHASH],
+        );
+
+        // Check that node two was not removed from the Kademlia table.
+        let two_removed_from_dht_table = check_for_line_with(
+            stdout1.clone(),
+            vec!["removed peer from kademlia table", SECP256K1MULTIHASH],
+        );
+
+        assert!(one_connected_to_two);
+        assert!(two_in_dht_routing_table);
+        assert!(two_added_to_dht);
+        assert!(two_disconnected_from_one);
+        assert!(!two_removed_from_dht_table);
+
+        // Check node two added node one to Kademlia table
+        let one_addded_to_dht = check_for_line_with(
+            stdout2.clone(),
+            vec![
+                "added configured node to kademlia routing table",
+                ED25519MULTIHASH,
+            ],
+        );
+
+        // Check node two DHT routing table was updated with node one
+        let one_in_dht_routing_table = check_for_line_with(
+            stdout2.clone(),
+            vec!["kademlia routing table updated with peer", ED25519MULTIHASH],
+        );
+
+        // Check that node two connected to node one.
+        let two_connected_to_one = check_for_line_with(
+            stdout2,
+            vec!["peer connection established", ED25519MULTIHASH],
+        );
+
+        assert!(one_addded_to_dht);
+        assert!(one_in_dht_routing_table);
+        assert!(two_connected_to_one);
+
+        // Check node endpoint to match
+        // let http_url = format!("http://localhost:{}", ws_port1);
+        // let http_resp = reqwest::get(format!("{}/node", http_url)).await.unwrap();
+        // assert_eq!(http_resp.status(), 200);
+        // let http_resp = http_resp.json::<serde_json::Value>().await.unwrap();
+        // assert_eq!(
+        //     http_resp,
+        //     serde_json::json!({
+        //             "static": {"peer_id": ED25519MULTIHASH},
+        //             "dynamic": {"listeners": [format!("{listen_addr1}")], "connections": {}}
+        //     })
+        // );
     });
 
     Ok(())
